requirements.txt
asteval==0.9.11
backports.ssl-match-hostname==3.5.0.1
certifi==2018.1.18
chardet==3.0.4
docker==3.0.1
docker-pycreds==0.2.2
get==0.0.39
idna==2.6
ipaddress==1.0.19
Jinja2==2.10
MarkupSafe==1.0
numpy==1.14.0
pkg-resources==0.0.0
post==0.0.26
public==0.0.65
query-string==0.0.28
request==0.0.26
requests==2.18.4
ruamel.ordereddict==0.4.13
ruamel.yaml==0.15.35
six==1.11.0
urllib3==1.22
websocket-client==0.46.0
